add src image in bd

+++ Herbal/app/src/main/java/com/example/herbal/DBHelper.java
@@ -15,23 +15,21 @@ import java.util.List;
 
     public class DBHelper extends SQLiteOpenHelper {
 
-        public static final int DATABASE_VERSION = 2;
+        public static final int DATABASE_VERSION = 8;
         public static final String DATABASE_NAME = "contactsDB";
         public static final String TABLE_CONTACTS = "contacts";
-    /*
+
         public static final String TABLE_IMAGES = "images";
-        public static final String TABLE_NOTE = "note";*/
+        public static final String TABLE_NOTE = "note";
 
         public static final String EXTERNAL_KEY_ID = "_id";
         public static final String CONTACTS_KEY_NAME = "name";
         public static final String CONTACTS_KEY_EMAIL = "email";
 
 
-       /* public static final String NOTE_KEY_TEXT = "text";
+        public static final String NOTE_KEY_TEXT = "text";
         public static final String NOTE_KEY_CREATEDATA = "data";
-
-        public static final String IMAGE_KEY_IMAGE = "blob";
-        public static final String IMAGE_FOREIGN_KEY_IDNOTE = "_idNote";*/
+        public static final String NOTE_KEY_IMAGE = "image";
 
 
        /* public DBHelper(Context context) {
@@ -42,63 +40,48 @@ import java.util.List;
         }
         @Override
         public void onCreate(SQLiteDatabase db) {
-            db.execSQL("create table " + TABLE_CONTACTS + "(" +
-                    EXTERNAL_KEY_ID + " integer primary key," +
-                    CONTACTS_KEY_NAME + " text," +
-                    CONTACTS_KEY_EMAIL + " text );");
-            //CreateTableNote(db);
-           // CreateTableImage(db);
+            Log.d("mLog", "-------------------onCreate---------------");
+            CreateContacts(db);
+            CreateTableNote(db);
         }
 
         @Override
         public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-            db.execSQL("drop table if exists " + TABLE_CONTACTS);
-            onCreate(db);
-           /* if (oldVersion != newVersion ){
-                DATABASE_VERSION = newVersion;
-                Log.d("mLog", " --- onUpgrade database from " + oldVersion
-                        + " to " + newVersion + " version --- ");
-                //db.execSQL("drop table if exists " + TABLE_CONTACTS);
-               // CreateContacts(db);
-                CreateTableNote(db);
-                CreateTableImage(db);
+            if (newVersion >  oldVersion){
 
+                    Log.d("mLog", oldVersion + " " + newVersion);
+                    Log.d("mLog", "-------------------onUpdate---------------");
+                    db.execSQL("drop table if exists " + TABLE_CONTACTS);
+                    db.execSQL("drop table if exists " + TABLE_NOTE);
+                    onCreate(db);
 
             }
-            else{
-               *//* db.execSQL("drop table if exists " + TABLE_CONTACTS);
-                db.execSQL("drop table if exists " + TABLE_IMAGES);
-                db.execSQL("drop table if exists " + TABLE_NOTE);
-                onCreate(db);*//*
-            }*/
 
         }
 
-        /*private void CreateNewTables(){
-
-        }
         private void CreateTableNote(SQLiteDatabase db){
             db.execSQL("create table " + TABLE_NOTE + "(" +
                     EXTERNAL_KEY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL," +
                     NOTE_KEY_TEXT + " TEXT," +
-                    NOTE_KEY_CREATEDATA + " data );");
+                    NOTE_KEY_IMAGE + " TEXT, " +
+                    NOTE_KEY_CREATEDATA + " text );");
             Log.d("mLog", " --- CREATE NOTE ---- ");
         }
-        private void CreateTableImage(SQLiteDatabase db){
+        /*private void CreateTableImage(SQLiteDatabase db){
             // создаем таблицу картинок
             db.execSQL("create table " + TABLE_IMAGES + "(" +
                     EXTERNAL_KEY_ID + " INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL," +
                     IMAGE_KEY_IMAGE + " blob," +
-                    IMAGE_FOREIGN_KEY_IDNOTE + " INTEGER," +
-                    "FOREIGN KEY (" + IMAGE_FOREIGN_KEY_IDNOTE + ") REFERENCES " + TABLE_NOTE + "(" + EXTERNAL_KEY_ID + ")" + " );");
+                    IMAGE_FOREIGN_KEY_IDNOTE + " INTEGER NOT NULL," +
+                    "FOREIGN KEY (" + IMAGE_FOREIGN_KEY_IDNOTE + ") REFERENCES " + TABLE_NOTE + "(" + EXTERNAL_KEY_ID + ")" + "ON UPDATE SET NULL" + " );");
             Log.d("mLog", " --- CREATE IMAGE ---- ");
-        }
+        }*/
         private void CreateContacts(SQLiteDatabase db){
             db.execSQL("create table " + TABLE_CONTACTS + "(" +
                     EXTERNAL_KEY_ID + " integer primary key," +
                     CONTACTS_KEY_NAME + " text," +
                     CONTACTS_KEY_EMAIL + " text );");
-        }*/
+        }
     }
